// deprecated-workflow-designer/sdc-workflow-designer-ui/src/app/model/node-data-type/node-data-type.ts
/**
 * Copyright (c) 2018 ZTE Corporation.
 * All rights reserved. This program and the accompanying materials
 * are made available under the terms of the Eclipse Public License v1.0
 * and the Apache License 2.0 which both accompany this distribution,
 * and are available at http://www.eclipse.org/legal/epl-v10.html
 * and http://www.apache.org/licenses/LICENSE-2.0
 *
 * Contributors:
 *     ZTE - initial API and implementation and/or initial documentation
 */
import { DisplayName } from "./display-name";
import { Icon } from "./icon";
import { SwaggerSchema } from "../workflow/swagger/swagger-schema";

export interface NodeDataType {
    id: string;
    displayName?: DisplayName;
    description?: DisplayName;
    type: string;
    icon: Icon;
    content: any;
    definitions: any;
}