// deprecated-workflow-designer/sdc-workflow-designer-ui/src/app/model/workflow/swagger/swagger-response.ts
/*******************************************************************************
 * Copyright (c) 2018 ZTE Corporation.
 * All rights reserved. This program and the accompanying materials
 * are made available under the terms of the Eclipse Public License v1.0
 * and the Apache License 2.0 which both accompany this distribution,
 * and are available at http://www.eclipse.org/legal/epl-v10.html
 * and http://www.apache.org/licenses/LICENSE-2.0
 *
 * Contributors:
 *     ZTE - initial API and implementation and/or initial documentation
 *******************************************************************************/
import { SwaggerSchema } from "./swagger-schema";

export interface SwaggerResponse {
    description: string;
    schema?: SwaggerSchema;
    headers?: any;//todo: add define object
    examples?: any;//todo: add define 
    $ref?: string;
    name?: string;
}